Refactor: Organize monolithic app.py into a modular structure
- **Blueprints**: I moved all routes from `app.py` into separate blueprint files within a new `routes/` directory. Each module (patients, admin, etc.) now has its own routing file.
- **Models**: I extracted all SQLAlchemy database models into a dedicated `models.py` file.
- **Extensions**: I centralized the SQLAlchemy instance in `extensions.py` to prevent circular imports.
- **Utilities & Decorators**: I moved shared helper functions and decorators into `utils.py` and `decorators.py` respectively.
- **Application Factory**: I refactored `app.py` to use the application factory pattern (`create_app`). The file is now responsible only for creating the app instance, registering blueprints, and running the application.

This new structure improves modularity, maintainability, and scalability, following standard Flask best practices.
